Move responsibilities from Run to Placement
Why:

A lot of queries / scopes on placements were being defined in Run.
Instead, they should be clearly-named scopes in Placement, accessed in
Run.

This clarifies the logic that's happening, and may make it easier to
test that the right placements are being accessed.

This is in preparation for more testing.

This change addresses the need by:

* Moving scopes to Placement (and one to Applicant).

* Debug #check_availability -- some placements in tests have 'nil'
automatic placements. Perhaps this should be a type coercion to integer,
instead of a NilCheck.