Debug Run prepopulation, availability tracking
Why:

Because of how arrays work in ActiveRecord and Postgres, without
explicitly marking the array as 'dirty', updating the availability
wasn't working.

Also, I had forgotten that after_initialize hooks run not only before
the object is saved for the first time, but every time the object is
read from the database. So, every time I re-read a Run object, it would
set all the positions to available.

This change addresses the need by:

* Change the after_initialize hook to an after_create hook, and update
the persistence accordingly.

* Mark #available_positions as 'dirty', or needing to be saved in
the database, when we update its contents.

* Also, we alter the interface to applicants and placements, for
simplicity elsewhere.