Add environment to migration.
Why:

When Matt (@allthesignals) was migrating, it was throwing an error that
Position wasn't available. This was because Run, which references
Position, wasn't being loaded in the run.rb file -- most of the time, it
just depends on the environment.

This change addresses the need by:

* Instead of explicitly requiring Position inside Run or in the
migration, we just load in the whole environment!

* Formatting the text to look a little better.