Update database tasks to use database URL
Why:

When trying to create the youth_match database on production, using a
DATABASE_URL environment variable, errors were being raised because the
config is a string, not a hash.

This change addresses the need by:

* Switching on case, handling databaseless configuration differently
based on whether it's a String or Hash.

* Moving the success message to inside the `begin` block, instead of
running after the error is rescued. This way, we only print 'created
database' when it's actually created.