Export mail merge to STDOUT, continue reorg
Why:

Heroku's filesystem is ephemeral, so it can't store export files.

This change addresses the need by:

* Changing the mail merge to output a string to STDOUT. We can use `>
filename` to divert to a file on the local system.

NOTE: We could probably do some conditional logic on, if not
environment, maybe environment variables like FILE=true or FILE=false.
Also, we need to do this for all the other file exports from production.

* Move the exports for allocation to the allocate rakefile.