Better errors, redirect, and pull-sync
Why:

This is a huge commit, in scope. Sorry.

First, we needed to refactor the Relay. We were getting so many errors
that weren't really errors as far as we can tell, but people clicking
extra times, possibly to change their decision.

Some people with errors have a "Self-Withdrawn" status. This is for
anyone who either withdrew from a job through the youth portal, or the
status that gets set when DYEE staff 'disassociate' people from jobs.
This counts as an 'expire', because most of these people are accidental
placements -- they were supposed to be CBO picks.

This change addresses the need by:

* Notify Airbrake for non-errors, and label them as such. Include much
more information in the data sent to Airbrake.

* Significantly refactor and clean up the logic in Relay.

* Set people with a 'self-withdrawn' status to expire.

* Rename #sync! to #push!, and add a #pull! method to pull in the
statuses that were manually changed in ICIMS before we start the next
round of placements.