Improve expiration
Why:

Jared found that records we had set to expired had not been set that way
in ICIMS. After looking, we discovered that the workflow was not being
updated in check_expiration.

It was difficult to understand the result of Workflow#status because it
was a code.

In order to make it easier for us to understand what the status is,

This change addresses the need by:

* Set the ICIMS workflow to expired when a record has expired.

* Remove applicant opt-out in loop, because Placement#check_expired
already does this.

* Add a status accessor that translates from code (e.g. "C#####") to
text (e.g. "accepted"), and update tests accordingly. Add icims_status
so that we can still get and set the code as necessary.

* Because unplacing could be dangerous, make the user confirm before
using the method.