Add offset to pulling.
Why:

Without this, pulling starts at 1. We've already pulled 150, so we don't
want to pull that first set again.

This change addresses the need by:

* Adding a param to cleanup rake task and query.

We should maybe have a :pulled boolean and a :pulled_at timestamp that
lets us know that we pulled someone's from ICIMS, and when.

That way we could query unpulled by pulled: false.