Expiration of broken workflows
Why:

Some workflows are nil or 0, so they can't be expired. ExpirationChecker
experienced an error because of this.

This change addresses the need by:

* Adding basically a nilcheck to Placement#check_expired.

NOTE: There should be a better way to do this, in Workflow or
Placement#workflow, which checks for bad workflows and returns a
NullObject / black hole.