Fix weighting
Why:

People were getting placed too far away -- the interest score was
outweighing travel time, so there were a lot of calls in to DYEE about
switching placements.

We want proximity to come first, then among the closest, maybe have the
interest score check

This change addresses the need by:

* Adding 12 to the travel score, to make it outweigh interest score.

NOTE: This is a quick-fix-style solution, but I wouldn't consider it a
"hack" in the traditional sense. In the future, I'd like to have the
lottery configuration include both a shift and a multiplier, and to use
that in the run instead.