Updates to allow for local setup
This update allows a local user to setup a database locally and run
the Youth Match algorithm on their local machine. It provides the
necessary files that can be imported into the local database for
doing the distance matching. It also appropriately buckets the
positions into automatic placements so they can all be automatically
placed. It also updates the README to give correct setup.